full cambios para csr

+++ app/Http/Controllers/CheckoutController.php
@@ -23,15 +23,15 @@ class CheckoutController extends Controller
         $items = [];
         $customer = auth()->guard('customer')->user();
         $cart = Cart::where('customer_id', $customer->id)->first();
-        if(!$cart) {
+        if (!$cart) {
             return redirect()->back()->with('error', 'No hay productos en el carrito');
         }
         $cartItems = $cart->items;
         $total = 0;
-        foreach($cartItems as $item) {
+        foreach ($cartItems as $item) {
             $price = $item->product->is_on_sale ? $item->product->sale_price : $item->product->price;
             $total += $price * $item->quantity;
-            $items [] = [
+            $items[] = [
                 'id' => $item->product->id,
                 'name' => $item->product->name,
                 'price' => $price,
@@ -48,12 +48,12 @@ class CheckoutController extends Controller
     {
         $delivery = Helper::getDeliveryCost();
         $product = Product::find($id);
-        if(!isset($product->stock->quantity)) {
+        if (!isset($product->stock->quantity)) {
             return redirect()->back()->with('error', 'Producto no disponible');
         }
         $price = $product->is_on_sale ? $product->sale_price : $product->price;
         $total = ($price * $quantity) + $delivery;
-        return view('frontend.checkout.checkout_direct', ['product' => $product, 'quantity' => $quantity, 'total' => $total,'price' => $price]);
+        return view('frontend.checkout.checkout_direct', ['product' => $product, 'quantity' => $quantity, 'total' => $total, 'price' => $price]);
     }
 
     /**
@@ -69,7 +69,8 @@ class CheckoutController extends Controller
      */
     public function store(Request $request)
     {
-        if($request->type == 'direct'){
+        $config = Helper::getConfigurations();
+        if ($request->type == 'direct') {
             $orderItems = [];
             $order = new Order();
             $order->customer_id = auth()->guard('customer')->user()->id;
@@ -87,7 +88,7 @@ class CheckoutController extends Controller
             $orderItem->quantity = 1;
             $orderItem->price = $request->total - Helper::getDeliveryCost();
             $orderItem->save();
-            $orderItems [] = [
+            $orderItems[] = [
                 'item_name' => $orderItem->product->name,
                 'quantity' => 1,
                 'price' => $request->total - Helper::getDeliveryCost(),
@@ -97,21 +98,44 @@ class CheckoutController extends Controller
             $product = Product::find($request->product_id);
             $product->stock->quantity -= 1;
             $product->stock->save();
-
             $customer_mail = $order->customer->email;
             $data_for_mail = [
                 'order' => $order,
                 'orderItem' => $orderItems,
             ];
-            Mail::to($customer_mail)->send(new OrderCreated($data_for_mail));
-            return response()->json([
-                'success' => true,
-                'redirect_url' => route('order.success', ['id' => $order->id])
-            ]);
+            if (env('PLAN') == 'basic') {
+                if ($config['whatsapp-number']) {
+                    $number = $config['whatsapp-number'];
+                    // Generar el mensaje de WhatsApp
+                    $message = "Hola, me gustaría hacer un pedido:\n\n";
+                    $message .= "Producto: {$orderItem->product->name}\n";
+                    $message .= "Precio: {$orderItem->price}\n";
+                    $message .= "Dirección de envío: {$order->address->full_address}\n";
+                    $message .= "Método de pago: {$order->paymentMethod->name}\n";
+                    $message .= "Total: {$order->total}\n";
+
+                    $whatsappUrl = "https://wa.me/{$number}?text=" . urlencode($message);
+
+                    return response()->json([
+                        'type' => 'whatsapp',
+                        'success' => true,
+                        'redirect_url' => $whatsappUrl
+                    ]);
+                }
+            }
+
+            if (env('PLAN') == 'medium' || env('PLAN') == 'pro') {
+                Mail::to($customer_mail)->send(new OrderCreated($data_for_mail));
+                return response()->json([
+                    'type' => 'email',
+                    'success' => true,
+                    'redirect_url' => route('order.success', ['id' => $order->id])
+                ]);
+            }
         }
         $customer = auth()->guard('customer')->user();
         $cart = Cart::where('customer_id', $customer->id)->first();
-        if(!$cart) {
+        if (!$cart) {
             return redirect()->back()->with('error', 'No hay productos en el carrito');
         }
         $cartItems = $cart->items;
@@ -125,7 +149,7 @@ class CheckoutController extends Controller
         $order->total = $request->total;
         $order->additional_charges = 0;
         $order->save();
-        foreach($cartItems as $item) {
+        foreach ($cartItems as $item) {
             $orderItem = new OrderItem();
             $orderItem->order_id = $order->id;
             $orderItem->product_id = $item->product_id;
@@ -133,7 +157,7 @@ class CheckoutController extends Controller
             $price = $item->product->is_on_sale ? $item->product->sale_price : $item->product->price;
             $orderItem->price = $price;
             $orderItem->save();
-            $orderItems [] = [
+            $orderItems[] = [
                 'item_name' => $orderItem->product->name,
                 'quantity' => $item->quantity,
                 'price' => $price,
@@ -149,13 +173,51 @@ class CheckoutController extends Controller
             'order' => $order,
             'orderItem' => $orderItems,
         ];
-        Mail::to($customer_mail)->send(new OrderCreated($data_for_mail));
-        $cart->items()->delete();
-        $cart->delete();
-        return response()->json([
-            'success' => true,
-            'redirect_url' => route('order.success', ['id' => $order->id])
-        ]);
+        if (env('PLAN') == 'basic') {
+            if ($config['whatsapp-number']) {
+                $number = $config['whatsapp-number'];
+                // Generar el mensaje de WhatsApp
+                $message = "Hola, me gustaría hacer un pedido:\n\n";
+                foreach ($cartItems as $item) {
+                    $message .= "Producto: {$item->product->name}\n";
+                    $price = $item->product->is_on_sale ? $item->product->sale_price : $item->product->price;
+                    $message .= "Precio: {$price}\n";
+                    $message .= "Cantidad: {$item->quantity}\n";
+                }
+                $message .= "Dirección de envío: {$order->address->full_address}\n";
+                $message .= "Método de pago: {$order->paymentMethod->name}\n";
+                $message .= "Total: {$order->total}\n";
+
+                $whatsappUrl = "https://wa.me/{$number}?text=" . urlencode($message);
+                $cart->items()->delete();
+                $cart->delete();
+                return response()->json([
+                    'type' => 'whatsapp',
+                    'success' => true,
+                    'redirect_url' => $whatsappUrl
+                ]);
+            }
+        }
+        if (env('PLAN') == 'medium') {
+            Mail::to($customer_mail)->send(new OrderCreated($data_for_mail));
+            $cart->items()->delete();
+            $cart->delete();
+            return response()->json([
+                'type' => 'email',
+                'success' => true,
+                'redirect_url' => route('order.success', ['id' => $order->id])
+            ]);
+        }
+        if (env('PLAN') == 'pro') {
+            Mail::to($customer_mail)->send(new OrderCreated($data_for_mail));
+            $cart->items()->delete();
+            $cart->delete();
+            return response()->json([
+                'type' => 'email',
+                'success' => true,
+                'redirect_url' => route('order.success', ['id' => $order->id])
+            ]);
+        }
     }
 
     /**
